Expire sandbox in unPersistSandboxSlice function
